# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aydın Campus Map API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/graph": {
      "get": {
        "summary": "Graph",
        "description": "The walking graph as GeoJSON (nodes and edges); ETag-cached.",
        "operationId": "graph_graph_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Graph Graph Get"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          }
        }
      }
    },
    "/places": {
      "get": {
        "summary": "Places",
        "description": "Word-prefix search; with a database, typos fall back to trigrams.\n\nWithout ``q``: the places the map shows (open areas and entrances at\nmeasured spots); rooms and doors known only from the tour are found\nby searching.",
        "operationId": "places_places_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceOut"
                  },
                  "title": "Response Places Places Get"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buildings": {
      "get": {
        "summary": "Buildings",
        "description": "Building massing summary (campus blocks by default).",
        "operationId": "buildings_buildings_get",
        "parameters": [
          {
            "name": "campus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Campus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildingOut"
                  },
                  "title": "Response Buildings Buildings Get"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/nodes/nearest": {
      "get": {
        "summary": "Nearest",
        "description": "Nearest outdoor/entrance panorama to a map click (equirectangular metres).\n\nIndoor spots stand at borrowed positions and are never the answer.",
        "operationId": "nearest_nodes_nearest_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Lng"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearestNode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/route": {
      "post": {
        "summary": "Route",
        "operationId": "route_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/weather": {
      "get": {
        "summary": "Weather",
        "operationId": "weather_weather_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Chat",
        "description": "Campus assistant; streams the Vercel AI SDK UI message protocol (v7).",
        "operationId": "chat_chat_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuildingOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "OSM way id"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "Campus block letter, e.g. A"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "campus": {
            "type": "boolean",
            "title": "Campus"
          },
          "height_m": {
            "type": "number",
            "title": "Height M"
          },
          "height_source": {
            "type": "string",
            "title": "Height Source",
            "description": "osm_levels | default | sfm"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "campus",
          "height_m",
          "height_source",
          "lng",
          "lat"
        ],
        "title": "BuildingOut"
      },
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalizedText": {
        "properties": {
          "tr": {
            "type": "string",
            "title": "Tr"
          },
          "en": {
            "type": "string",
            "title": "En"
          }
        },
        "type": "object",
        "required": [
          "tr",
          "en"
        ],
        "title": "LocalizedText"
      },
      "NearestNode": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "distance_m": {
            "type": "number",
            "title": "Distance M"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "distance_m"
        ],
        "title": "NearestNode"
      },
      "NodeKind": {
        "type": "string",
        "enum": [
          "outdoor",
          "entrance",
          "indoor"
        ],
        "title": "NodeKind"
      },
      "PlaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Node id of the representative panorama"
          },
          "name_tr": {
            "type": "string",
            "title": "Name Tr"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "kind": {
            "$ref": "#/components/schemas/NodeKind"
          },
          "building": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building"
          },
          "node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Node Ids"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor",
            "description": "Storey of a room (kind indoor), when known"
          },
          "area_tr": {
            "type": "string",
            "title": "Area Tr",
            "description": "Tour area, e.g. Kimya - M Blok",
            "default": ""
          },
          "area_en": {
            "type": "string",
            "title": "Area En",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name_tr",
          "name_en",
          "kind",
          "building",
          "node_ids"
        ],
        "title": "PlaceOut"
      },
      "RouteRequest": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Start node id (a place id is a node id)"
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Destination node id"
          },
          "avoid_stairs": {
            "type": "boolean",
            "title": "Avoid Stairs",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "RouteRequest"
      },
      "RouteResponse": {
        "properties": {
          "node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Node Ids"
          },
          "length_m": {
            "type": "number",
            "title": "Length M"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "coordinates": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Coordinates",
            "description": "[lng, lat] per node; indoor spots repeat their entrance's position: draw `parts`, not this"
          },
          "parts": {
            "items": {
              "items": {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              "type": "array"
            },
            "type": "array",
            "title": "Parts",
            "description": "[lng, lat] runs the map can draw: measured nodes joined by walks outside (the line breaks at indoor spots and passages)"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "type": "array",
            "title": "Steps"
          },
          "approximate": {
            "type": "boolean",
            "title": "Approximate",
            "description": "Length and duration include estimated indoor stretches",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "node_ids",
          "length_m",
          "duration_s",
          "coordinates",
          "steps"
        ],
        "title": "RouteResponse"
      },
      "StepOut": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "turn": {
            "$ref": "#/components/schemas/Turn"
          },
          "distance_m": {
            "type": "number",
            "title": "Distance M",
            "description": "Walked after this step (estimate indoors)"
          },
          "bearing_deg": {
            "type": "number",
            "title": "Bearing Deg",
            "description": "Compass bearing outdoors; 0 indoors"
          },
          "text_tr": {
            "type": "string",
            "title": "Text Tr"
          },
          "text_en": {
            "type": "string",
            "title": "Text En"
          },
          "indoor": {
            "type": "boolean",
            "title": "Indoor",
            "description": "Inside a building: no compass, the distance is an estimate",
            "default": false
          },
          "building": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building",
            "description": "Block the step is in, enters or leaves"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor",
            "description": "Storey at the step's spot (0 = ground)"
          },
          "floors": {
            "type": "integer",
            "title": "Floors",
            "description": "Storeys climbed (+) or descended (-) by stairs",
            "default": 0
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedText"
              },
              {
                "type": "null"
              }
            ],
            "description": "Indoor spot walked to (or started at)"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "turn",
          "distance_m",
          "bearing_deg",
          "text_tr",
          "text_en"
        ],
        "title": "StepOut"
      },
      "Turn": {
        "type": "string",
        "enum": [
          "start",
          "straight",
          "slight_left",
          "left",
          "sharp_left",
          "slight_right",
          "right",
          "sharp_right",
          "u_turn",
          "arrive",
          "enter",
          "exit",
          "go_to",
          "stairs_up",
          "stairs_down",
          "through"
        ],
        "title": "Turn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeatherOut": {
        "properties": {
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "temperature_c": {
            "type": "number",
            "title": "Temperature C"
          },
          "weather_code": {
            "type": "integer",
            "title": "Weather Code",
            "description": "WMO weather interpretation code"
          },
          "cloud_cover_pct": {
            "type": "number",
            "title": "Cloud Cover Pct"
          },
          "precipitation_mm": {
            "type": "number",
            "title": "Precipitation Mm"
          },
          "wind_speed_kmh": {
            "type": "number",
            "title": "Wind Speed Kmh"
          },
          "wind_direction_deg": {
            "type": "number",
            "title": "Wind Direction Deg"
          },
          "visibility_m": {
            "type": "number",
            "title": "Visibility M"
          },
          "is_day": {
            "type": "boolean",
            "title": "Is Day"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "attribution": {
            "type": "string",
            "title": "Attribution"
          }
        },
        "type": "object",
        "required": [
          "observed_at",
          "temperature_c",
          "weather_code",
          "cloud_cover_pct",
          "precipitation_mm",
          "wind_speed_kmh",
          "wind_direction_deg",
          "visibility_m",
          "is_day",
          "stale",
          "attribution"
        ],
        "title": "WeatherOut"
      }
    }
  }
}